requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asgi-lifespan==2.1.0
asyncpg==0.28.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
filelock==3.12.2
Flask==2.3.2
Flask-API==3.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
identify==2.5.27
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy==1.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
passlib==1.7.4
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlalchemy-stubs==0.4
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
tomli==2.0.1
types-click==7.1.8
types-Flask==1.1.6
types-Flask-Migrate==4.0.0.6
types-Flask-SQLAlchemy==2.5.9.4
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-passlib==1.7.7.13
types-pyasn1==0.4.0.6
types-python-jose==3.3.4.8
types-requests==2.31.0.2
types-SQLAlchemy==1.4.53.38
types-urllib3==1.26.25.14
types-Werkzeug==1.0.9
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.3
Werkzeug==2.3.7
WTForms==3.0.1

a2wsgi~=1.7.0
